Add a limitation section because pb in interrup
Dear @mobizt , first congratulations for your very extensive email lib ;-)
I would like to add this section ("Limitation"), because I spent lot of time to try to send an email in an "Interrup", and my ESP32 crash systematicaly when sending email. My use case is a PIR sensor that start an interupt. You can rephrase or change totaly this content ;-), but the idea is to inform your user that they will have an hard-time if the try to send email in interrup (like me !)
Regards,
Nicolas.